pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "text2speech"
version = "0.2.0"
description = "Provides text2speech capabilities using ElevenLabs and Kokoro TTS."
authors = [
    { name = "Daniel Gaida", email = "daniel.gaida@th-koeln.de" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.8"

dependencies = [
    "torch>=2.0.0,<3.0.0",
    "torchaudio>=2.0.0,<3.0.0",
    "sounddevice>=0.4.0,<0.5.0",
    "pyyaml>=6.0.0,<7.0.0",
    "kokoro>=0.1.0,<1.2.0",
    "elevenlabs>=1.0.0",
    "cachetools>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0,<8.0.0",
    "pytest-cov>=4.0.0,<5.0.0",
    "pytest-mock>=3.10.0,<4.0.0",
    "pytest-timeout>=2.1.0,<3.0.0",
    "ruff>=0.7.0,<1.0.0",
    "black>=24.0.0,<25.0.0",
    "mypy>=1.0.0,<2.0.0",
    "bandit[toml]>=1.7.0,<2.0.0",
    "pre-commit>=3.0.0,<4.0.0",
    "build>=1.0.0,<2.0.0",
    "twine>=5.0.0,<6.0.0",
    "hypothesis>=6.0.0",
    "types-PyYAML",
    "types-cachetools",
]
test = [
    "pytest>=7.0.0,<8.0.0",
    "pytest-cov>=4.0.0,<5.0.0",
    "pytest-mock>=3.10.0,<4.0.0",
    "pytest-timeout>=2.1.0,<3.0.0",
    "hypothesis>=6.0.0",
]

[project.scripts]
text2speech = "text2speech.cli:main"

[project.urls]
Homepage = "https://github.com/dgaida/text2speech"
Repository = "https://github.com/dgaida/text2speech"

[tool.setuptools]
packages = ["text2speech"]

[tool.setuptools.package-data]
text2speech = ["py.typed"]

[tool.black]
line-length = 127
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | __pycache__
)/
'''

[tool.ruff]
line-length = 127
target-version = "py38"

[tool.bandit]
exclude_dirs = ["tests/", "scripts/", "examples/"]
skips = ["B101"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
# Exclude examples directory from test collection
norecursedirs = ["examples", ".git", ".tox", "dist", "build", "*.egg", "scripts"]
addopts = [
    "-v",
    "--strict-markers",
    "--strict-config",
]
markers = [
    "integration: marks tests as integration tests",
]
